+++ frontend/src/views/sql-editor/EditorPanel/EditorPanel.vue
@@ -1,60 +1,35 @@
 <template>
   <div class="w-full flex-1 flex flex-row items-stretch overflow-hidden">
-    <Splitpanes
-      v-if="
-        !currentTab ||
-        currentTab.mode === 'READONLY' ||
-        currentTab.mode === 'STANDARD'
-      "
-      horizontal
-      class="default-theme"
-      :dbl-click-splitter="false"
-    >
-      <Pane class="flex flex-row overflow-hidden">
-        <StandardPanel v-if="isDisconnected || allowReadonlyMode" />
-        <ReadonlyModeNotSupported v-else />
-      </Pane>
-      <Pane
-        v-if="!isDisconnected && allowReadonlyMode"
-        class="relative"
-        :size="40"
-      >
-        <ResultPanel />
-      </Pane>
-    </Splitpanes>
+    <Panels>
+      <template #code-panel>
+        <StandardPanel
+          v-if="
+            !currentTab ||
+            currentTab.mode === 'READONLY' ||
+            currentTab.mode === 'STANDARD'
+          "
+        />
 
-    <TerminalPanel v-else-if="currentTab.mode === 'ADMIN'" />
+        <TerminalPanel v-else-if="currentTab.mode === 'ADMIN'" />
 
-    <AccessDenied v-else />
+        <AccessDenied v-else />
+      </template>
+    </Panels>
   </div>
 </template>
 
 <script setup lang="ts">
 import { storeToRefs } from "pinia";
-import { Pane, Splitpanes } from "splitpanes";
-import { computed } from "vue";
-import {
-  useConnectionOfCurrentSQLEditorTab,
-  useSQLEditorTabStore,
-} from "@/store";
-import { instanceV1HasReadonlyMode } from "@/utils";
+import { useSQLEditorTabStore } from "@/store";
 import AccessDenied from "./AccessDenied.vue";
-import ReadonlyModeNotSupported from "./ReadonlyModeNotSupported.vue";
-import ResultPanel from "./ResultPanel";
+import Panels from "./Panels";
 import StandardPanel from "./StandardPanel";
 import TerminalPanel from "./TerminalPanel";
 import { provideEditorPanelContext } from "./context";
 
 const tabStore = useSQLEditorTabStore();
-const { currentTab, isDisconnected } = storeToRefs(tabStore);
-const { instance } = useConnectionOfCurrentSQLEditorTab();
+const { currentTab } = storeToRefs(tabStore);
 provideEditorPanelContext({
   tab: currentTab,
-});
-
-const allowReadonlyMode = computed(() => {
-  if (isDisconnected.value) return false;
-
-  return instanceV1HasReadonlyMode(instance.value);
 });
 </script>
